+++ 1-foundation/practical-logger-core/CHANGELOG.md
@@ -6,6 +6,7 @@
 - TODO assess the arguments normalization
 - TODO link to other modules
 - [test] improved demo ~ "unit test"
+* [feat] new API: group() groupCollapsed() groupEnd()
 
 ## v1.0.0
 2019/04/04