+++ 1-foundation/practical-logger-node/CHANGELOG.md
@@ -7,6 +7,7 @@
 * [chore] bumped deps
 * [chore] tslib no longer a peerDependency
 - [test] improved demo ~ "unit test"
+* [feat] new API: group() groupCollapsed() groupEnd()
 
 ## v0.0.1
 2019/04/02